probe: Move files from tools package into probe.
"tools" has no particular meaning, so I move the files into an
easy-to-understand package called "probe".
